individual post view working, delete works now too
