+++ .gitignore
@@ -15,3 +15,4 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.env
