+++ client/src/Pages/Dallas.js
@@ -1,234 +1,39 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
+import DallasForm from "../components/DallasForm";
 import DallasPostList from "../components/DallasPostList";
+import IndividualPost from "../components/IndividualPost";
 
-const Dallas = (props) => {
-  // Initial post in case that you want to update a new post
-  const {
-    initialPost = {
-      id: null,
-      username: "",
-      imageurl: "",
-      alt: "",
-      dish: "",
-      restaurant: "",
-      content: "",
-      alias: "",
-      date: "",
-    },
-  } = props;
+const Dallas = () => {
+  const [selected, setSelected] = useState(null);
+  //const [posts, setPosts] = useState([]);
 
-  // We're using that initial student as our initial state
-  const [post, setPost] = useState(initialPost);
-
-  const [matchingRestaurants, setMatchingRestaurants] = useState([]);
-
-  //create functions that handle the event of the user typing into the form
-  const handleUsernameChange = (event) => {
-    const username = event.target.value;
-    setPost((post) => ({ ...post, username }));
-  };
-
-  const handleImageURLChange = (event) => {
-    const imageurl = event.target.value;
-    setPost((post) => ({ ...post, imageurl }));
-  };
-
-  const handleAltChange = (event) => {
-    const alt = event.target.value;
-    setPost((post) => ({ ...post, alt }));
-  };
-
-  const handleDishChange = (event) => {
-    const dish = event.target.value;
-    setPost((post) => ({ ...post, dish }));
-  };
-
-  const handleRestaurantChange = (event) => {
-    const restaurantQuery = event.target.value;
-    getRestaurants(event, restaurantQuery);
-    setPost((post) => ({ ...post, restaurant: restaurantQuery }));
-    //setPost("");
-    //event.target.value = showRestaurant;
-  };
-
-  const handleContentChange = (event) => {
-    const content = event.target.value;
-    setPost((post) => ({ ...post, content }));
-  };
-
-  const handleAliasChange = (event) => {
-    const alias = event.target.value;
-    setPost((post) => ({ ...post, alias }));
-  };
-
-  const handleDateChange = (event) => {
-    const date = event.target.value;
-    setPost((post) => ({ ...post, date }));
-  };
-
-  //getting the API data from the server
-  function getRestaurants(event, restaurant) {
-    event.preventDefault();
-    //restaurant
-    fetch(`/api/location-search?term=${restaurant}`, {
-      method: "GET",
-      headers: { "Content-Type": "application/json" },
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        console.log("data check line 76 ", data.businesses);
-        let matchingRestaurantsTemp = data.businesses.map((business) => ({
-          alias: business.alias,
-          name: business.name,
-          address: business.location.display_address,
-        }));
-        setMatchingRestaurants(matchingRestaurantsTemp);
-        console.log("checking fetchRestaurant ", matchingRestaurantsTemp);
-        console.log("checking restaurants ", restaurant);
-      })
-      .catch((err) => console.error(err));
+  function passingSelected(post) {
+    //console.log('hello')
+    setSelected(post);
+    console.log("post dallas check", post);
   }
 
-  //A function to handle the post request
-  const newPosts = async (newPost) => {
-    const response = await fetch("/api/blogposts", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(newPost),
-    });
-    const data = await response.json();
-    console.log("From the post ", data);
-    props.savePost(data);
-  };
+  let switchDisplayView;
 
-  //a function to handle the Update request
-  const updatePost = (existingPost) => {
-    return fetch(`/api/blogposts/${existingPost.id}`, {
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(existingPost),
-    })
-      .then((response) => {
-        return response.json();
-      })
-      .then((data) => {
-        console.log("From put request ", data);
-        props.savePost(data);
-      });
-  };
-
-  // Than handle submit function now needs the logic for the update scenario
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (post.id) {
-      updatePost(post);
-    } else {
-      newPosts(post);
-    }
-  };
-
-  function chosenRestaurant(restaurant) {
-    let input = restaurant;
-    setPost((post) => ({
-      ...post,
-      restaurant: restaurant.name,
-      alias: restaurant.alias,
-    }));
-    console.log("inputCheck", input);
-  }
-
-  return (
-    <div>
-      <h1>Dallas Ventures Heeere</h1>
-
-      <form onSubmit={handleSubmit}>
-        <fieldset>
-          <label>Username</label>
-          <input
-            type="text"
-            id="add-user-name"
-            placeholder="Username"
-            required
-            value={post.username}
-            onChange={handleUsernameChange}
-          />
-          <label>Image URL</label>
-          <input
-            type="text"
-            id="add-image-url"
-            placeholder="Image URL"
-            required
-            value={post.imageurl}
-            onChange={handleImageURLChange}
-          />
-          <label>Image Description</label>
-          <input
-            type="text"
-            id="add-alt"
-            placeholder="Alt"
-            required
-            value={post.alt}
-            onChange={handleAltChange}
-          />
-          <label>Dish Name</label>
-          <input
-            type="text"
-            id="add-dish-name"
-            placeholder="Dish Name"
-            required
-            value={post.dish}
-            onChange={handleDishChange}
-          />
-          <label>Restaurant Name</label>
-          <input
-            type="text"
-            id="add-restaurant"
-            placeholder="Restaurant Name"
-            required
-            value={post.restaurant}
-            onChange={handleRestaurantChange}
-          />
-
-          <input type="hidden" id="alias" required value={post.alias} />
-          <label>Content</label>
-          <input
-            type="text"
-            id="add-content"
-            placeholder="Content Here"
-            required
-            value={post.content}
-            onChange={handleContentChange}
-          />
-          <label>Date</label>
-          <input
-            type="date"
-            id="add-date"
-            placeholder="Date"
-            required
-            value={post.date}
-            onChange={handleDateChange}
-          />
-        </fieldset>
-        <button type="submit">{!post.id ? "Submit" : "Save"}</button> <br />
-      </form>
-
+  if (selected != null) {
+    switchDisplayView = (
       <div>
-        {matchingRestaurants.map((restaurant) => (
-          <ul key={restaurant.id}>
-            <li>
-              {" "}
-              <a onClick={() => chosenRestaurant(restaurant)}>
-                {" "}
-                {restaurant.name} {JSON.stringify(restaurant)}{" "}
-              </a>{" "}
-            </li>
-          </ul>
-        ))}
-      </div>
+        <IndividualPost post={selected} />
+        <br />
 
-      <DallasPostList />
-    </div>
-  );
+        <button onClick={() => setSelected(null)}>Go Back</button>
+      </div>
+    );
+  } else {
+    switchDisplayView = (
+      <div>
+        <DallasForm />
+        <DallasPostList passingSelected={passingSelected} />
+      </div>
+    );
+  }
+
+  return <div>{switchDisplayView}</div>;
 };
 
 export default Dallas;
